Add Autoloader for keys and environment variables
This is another lengthy change that adds support for automating the
loading of ssh and gpg keys to their respective agents and creating
environment files for loading into the shell.

Whilst additional effort has been placed into re-working much of the api
and building additional tests, this should now result in smaller, more
targetted commits that reduce the overall overhead.

This change is necessary to allow the autoloading of keys directly from
bitwarden vault to the respective `ssh` and `gpg` agents-

This change addresses the need by: introducing an autoloader

As a result of this change, in local testing the following side effects
were observed:

- some tests currently fail