Fix user login session key store
As part of the change to implementing the autoloader, a goroutine was
introduced that wrapped setup and added the user hash password as a
session key. This caused a null pointer panic as the user cache was not
populated before the cache was populated for the first time.

This change is necessary to resolve panics whilst using user auth

This change addresses the need by:

- Introducing an additional channel to the refresh method which is
  written to once the first sync is complete. This blocks the caller
  preventing it from continuing until we know the cache is fully
  populated.

As a result of this change, in local testing the following side effects
were observed:

- Broken tests were fixed.